refactor calls query out
