+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallsPage/callsTableQuery.tsx
@@ -1,0 +1,38 @@
+import {useMemo} from 'react';
+
+import {useWFHooks} from '../wfReactInterface/context';
+import {CallFilter} from '../wfReactInterface/wfDataModelHooksInterface';
+import {WFHighLevelCallFilter} from './callsTableFilter';
+
+/**
+ * Given a calls query, resolve the calls and return them. This hook will
+ * serve as an in-memory stepping stone to a more sophisticated backend
+ * implementation.
+ */
+export const useCallsForQuery = (
+  entity: string,
+  project: string,
+  filter: WFHighLevelCallFilter
+) => {
+  const {useCalls} = useWFHooks();
+  const lowLevelFilter: CallFilter = useMemo(() => {
+    return convertHighLevelFilterToLowLevelFilter(filter);
+  }, [filter]);
+
+  const calls = useCalls(entity, project, lowLevelFilter);
+  return calls;
+};
+
+const convertHighLevelFilterToLowLevelFilter = (
+  effectiveFilter: WFHighLevelCallFilter
+): CallFilter => {
+  return {
+    traceRootsOnly: effectiveFilter.traceRootsOnly,
+    opVersionRefs: effectiveFilter.opVersionRefs,
+    inputObjectVersionRefs: effectiveFilter.inputObjectVersionRefs,
+    outputObjectVersionRefs: effectiveFilter.outputObjectVersionRefs,
+    parentIds: effectiveFilter.parentId
+      ? [effectiveFilter.parentId]
+      : undefined,
+  };
+};
